brute_protected_auth/requirements.txt
Django==2.2.3
djangorestframework==3.9.4
psycopg2-binary==2.8.3
pytz==2019.1
redis==3.4.1
sqlparse==0.3.0
